# backend/src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: API记录系统
  description: 用于记录和管理API调用信息的系统
  version: 1.0.0
  contact:
    name: 开发团队
    email: dev@example.com
servers:
  - url: http://localhost:8080
    description: 开发环境
  - url: http://172.16.2.95:8080
    description: 生产环境
paths:
...
  /recorder/records:
    post:
      summary: 记录API调用
      description: 记录API调用的详细信息
      tags:
        - API记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiRecordDto'
      responses:
        '201':
          description: API调用记录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiRecordDto'
        '400':
          description: 请求参数错误

    get:
      summary: 获取API调用记录
      description: 分页获取API调用记录列表
      tags:
        - API记录
      security:
        - BasicAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: 页码（从0开始）
        - name: size
          in: query
          schema:
            type: integer
            default: 10
          description: 每页大小
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: 开始日期（YYYY-MM-DD）
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: 结束日期（YYYY-MM-DD）
      responses:
        '200':
          description: 成功获取API调用记录
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiRecordDto'
                  totalElements:
                    type: integer
                    description: 总记录数
                  totalPages:
                    type: integer
                    description: 总页数
                  size:
                    type: integer
                    description: 每页大小
                  number:
                    type: integer
                    description: 当前页码
        '401':
          description: 未认证
  /recorder/stats:
    get:
      summary: 获取API调用统计
      description: 获取API调用的统计信息
      tags:
        - API记录
      security:
        - BasicAuth: []
      responses:
        '200':
          description: 成功获取统计信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCalls:
                    type: integer
                    description: 总调用次数
                  todayCalls:
                    type: integer
                    description: 今日调用次数
                  successCalls:
                    type: integer
                    description: 成功调用次数
                  errorCalls:
                    type: integer
                    description: 错误调用次数
        '401':
          description: 未认证
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  schemas:
...
    ApiRecordDto:
      type: object
      description: API调用记录对象
      required:
        - url
        - method
        - callTime
      properties:
        id:
          type: integer
          description: 记录ID
        url:
          type: string
          description: 请求URL
          example: "http://example.com/api/users"
        method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH]
          description: HTTP方法
          example: "POST"
        headers:
          type: object
          description: 请求头信息
          additionalProperties: true
        body:
          type: string
          description: 请求体内容
        callTime:
          type: string
          format: date-time
          description: 调用时间
        responseStatus:
          type: integer
          description: 响应状态码
          example: 200
        responseBody:
          type: string
          description: 响应体内容
        createdAt:
          type: string
          format: date-time
          description: 创建时间
...